third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-childNodeCount.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function addNode()
{
    var container = document.getElementById("container");
    container.appendChild(document.createElement("div"));
}

function removeNode()
{
    var container = document.getElementById("container");
    container.firstChild.remove();
}

function test()
{
    var nodeInfo = {};
    var containerNodeId;
    InspectorTest.eventHandler["DOM.setChildNodes"] = setChildNodes;
    InspectorTest.eventHandler["DOM.childNodeCountUpdated"] = childNodeCountUpdated;

    InspectorTest.sendCommand("DOM.getDocument", {}, onGotDocument);

    function onGotDocument(msg)
    {
        if (InspectorTest.completeTestIfError(msg))
            return;
        InspectorTest.sendCommand("DOM.querySelector", { nodeId: msg.result.root.nodeId, selector: "#container" }, onQuerySelector);
    }

    function onQuerySelector(msg)
    {
        if (InspectorTest.completeTestIfError(msg))
            return;
        containerNodeId = msg.result.nodeId;
        InspectorTest.log("Node arrived with childNodeCount: " + nodeInfo[containerNodeId].childNodeCount);

        InspectorTest.sendCommand("Runtime.evaluate", { expression: "addNode()"});
        InspectorTest.sendCommand("Runtime.evaluate", { expression: "removeNode()"});
        InspectorTest.sendCommand("Runtime.evaluate", { expression: "removeNode()"});
        InspectorTest.sendCommand("Runtime.evaluate", { expression: "removeNode()"},
            InspectorTest.completeTest.bind(InspectorTest));
    }

    function setChildNodes(message)
    {
        var nodes = message.params.nodes;
        for (var i = 0; i < nodes.length; ++i) {
            nodeInfo[nodes[i].nodeId] = nodes[i];
            delete nodes[i].nodeId;
        }
    }

    function childNodeCountUpdated(message)
    {
        if (message.params.nodeId === containerNodeId)
            InspectorTest.log("childCountUpdated: " + message.params.childNodeCount);
    }
}

</script>
</head>
<body onload="runTest()">
<div id="container" style="display:none"><div>child1</div><div>child2</div></div>
</body>
</html>
