third_party/blink/web_tests/platform/generic/inspector-protocol/dom/dom-childNodeCount-expected.txt

Tests how DOM.childNodeCountUpdated event works.
Node arrived with childNodeCount: 2
childCountUpdated: 3
childCountUpdated: 2
childCountUpdated: 1
childCountUpdated: 0
